Restructure auth to prevent next-auth type errors
