+++ app/components/Lobby.tsx
@@ -2,7 +2,7 @@ import { redirect } from "next/navigation";
 import Image from "next/image";
 import Link from "next/link";
 import { getServerSession } from "next-auth";
-import { authOptions } from "../api/auth/[...nextauth]/route";
+import { authOptions } from "../api/auth/[...nextauth]/config";
 import { Title } from "./Title";
 import { DarkModeToggle } from "./DarkThemeToggle";
 import { LobbyForm } from "./forms/LobbyForm";
