+++ app/play/[[...slug]]/page.tsx
@@ -1,5 +1,5 @@
 import { getServerSession } from "next-auth";
-import { authOptions } from "../../api/auth/[...nextauth]/route";
+import { authOptions } from "../../api/auth/[...nextauth]/config";
 import { PageProps } from "@/.next/types/app/layout";
 import { redirect } from "next/navigation";
 import Link from "next/link";
